+++ src/server/common/helpers/upload-status.js
