+++ src/server/upload/controller.js
@@ -7,7 +7,7 @@ import {
 import {
   uploadStatus,
   canBeQuarantined
-} from '~/src/server/common/upload-status'
+} from '~/src/server/common/helpers/upload-status'
 import {
   storeUploadDetails,
   findUploadDetails
